database name changed, routes and controllers created, answer model to Answer, user favorites fixed, changed app import of routes
