+++ api/src/app.js
@@ -2,7 +2,7 @@ const express = require('express');
 const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
 const morgan = require('morgan');
-const routes = require('./routes/Routes.js');
+const routes = require('./routes/index');
 
 require('./db.js');
