+++ api/src/controllers/answerControllers.js
@@ -1,0 +1,18 @@
+const postAnswer=(req,res,next)=>{
+    //user.findbypk=usuario que respondio=>usuario.addAnswer(respuesta crada)/user.findbypk=question=>Question.addAnswer(respuesta crada)
+    //llega por body respuesta, id de pregunta y id de usuario=>añade respuesta a usuario, pregunta, cdrea respuesta
+}
+
+const putAnswer=(req,res,next)=>{
+   
+}
+
+const deleteAnswer=(req,res,next)=>{
+   
+}
+
+module.exports={
+    postAnswer,
+    putAnswer,
+    deleteAnswer,
+}
