+++ api/src/controllers/answersControllers.js
@@ -1,0 +1,8 @@
+const getAnswers=(req,res,next)=>{
+   
+}
+
+
+module.exports={
+  getAnswers
+}
