+++ api/src/controllers/usersControllers.js
@@ -1,0 +1,12 @@
+const getTopTen= (req,res,next)=>{
+    
+}
+
+const getRanking=(req,res,next)=>{
+    //  /users?sort=ascendent/descendent
+ }
+
+module.exports={
+    getTopTen,
+    getRanking
+}
